Support for various score types and improved code
Refactored the code, so that the score type is now an argument of the
command-line call. Currently supported are: av, l5, l3, av5, av3.
The scores are then computed from the raw score-confidence data and
statistics are printed for those particular scores. To compare statistics for
difference score types, just run the program multiple times with different
arguments. It's easy to extend the code by additional score types now.