Overlap, R-Test, Correlation, better documentation
1. Add option --overlap to output overlap for all acceptance rates
2. Add option --rtest that computes the p-value for a sampled R-test
3. Now output correlation instead of distance for Kendall tau
4. Improved the documentation in various places

TODO: There is now a bit of code duplication because --overlap also
creates a gnuplot script, just like --print

TODO: Change the Kendall tau functions to compute the correlation
directly. It's weird to do work there to compute a distance, only to
later convert it to a correlation again. The relation is trivial anyway,
and the correlation is the more natural measure for Kendall's tau.

TODO: Usage info is too long now because of the many scores types. Only
show the most important score types and hide the others behind an
additional options like --show-score-types.